# mikroSDK `v2.3.0` ## Changes
### TIVA SDK

Support added for 71 Texas Instruments TIVA MCU's.

| CHIP NAME       | CHIP NAME       | CHIP NAME       |
|-----------------|-----------------|-----------------|
| TM4C1230C3PM    | TM4C1230D5PM    | TM4C1230E6PM    |
| TM4C1230H6PM    | TM4C1231C3PM    | TM4C1231D5PM    |
| TM4C1231D5PZ    | TM4C1231E6PM    | TM4C1231E6PZ    |
| TM4C1231H6PGE   | TM4C1231H6PM    | TM4C1231H6PZ    |
| TM4C1232C3PM    | TM4C1232D5PM    | TM4C1232E6PM    |
| TM4C1232H6PM    | TM4C1233C3PM    | TM4C1233D5PM    |
| TM4C1233D5PZ    | TM4C1233E6PM    | TM4C1233E6PZ    |
| TM4C1233H6PGE   | TM4C1233H6PM    | TM4C1233H6PZ    |
| TM4C1236D5PM    | TM4C1236E6PM    | TM4C1236H6PM    |
| TM4C1237D5PM    | TM4C1237D5PZ    | TM4C1237E6PM    |
| TM4C1237E6PZ    | TM4C1237H6PGE   | TM4C1237H6PM    |
| TM4C1237H6PZ    | TM4C123AE6PM    | TM4C123AH6PM    |
| TM4C123BE6PM    | TM4C123BE6PZ    | TM4C123BH6PGE   |
| TM4C123BH6PM    | TM4C123BH6PZ    | TM4C123BH6ZRB   |
| TM4C123FE6PM    | TM4C123FH6PM    | TM4C123GE6PM    |
| TM4C123GE6PZ    | TM4C123GH6PGE   | TM4C123GH6PM    |
| TM4C123GH6PZ    | TM4C123GH6ZRB   | TM4C123GH6ZXR   |
| TM4C1290NCPDT   | TM4C1290NCZAD   | TM4C1292NCPDT   |
| TM4C1292NCZAD   | TM4C1294KCPDT   | TM4C1294NCPDT   |
| TM4C1294NCZAD   | TM4C1297NCZAD   | TM4C1299KCZAD   |
| TM4C1299NCZAD   | TM4C129CNCPDT   | TM4C129CNCZAD   |
| TM4C129DNCPDT   | TM4C129DNCZAD   | TM4C129EKCPDT   |
| TM4C129ENCPDT   | TM4C129ENCZAD   | TM4C129LNCZAD   |
| TM4C129XKCZAD   | TM4C129XNCZAD   |

Supported modules:

1. ADC
   + Full module support
2. SPI
   + Full module support
3. UART
   + Full module support
   + Implemented using interrupts
4. I2C
   + Full module support
5. PWM
   + Full module support
6. GPIO
   + Full module support

### NEW HARDWARE

+ New board definitions added:
  + Fusion for TIVA v8 - <https://www.mikroe.com/fusion-for-tiva>
  + EasyMX Pro v7 for TIVA - <https://www.mikroe.com/easymx-pro-tiva>
  + MCU CARD FOR TIVA TM4C129XNCZAD - <https://www.mikroe.com/mcu-card-for-tiva-tm4c129xnczad>
  + MCU CARD FOR TIVA TM4C1294NCZAD - <https://www.mikroe.com/mcu-card-for-tiva-tm4c1294nczad>
  + MCU CARD FOR TIVA TM4C129XKCZAD - <https://www.mikroe.com/mcu-card-for-tiva-tm4c129xkczad>
  + MCU CARD FOR TIVA TM4C129LNCZAD - <https://www.mikroe.com/mcu-card-for-tiva-tm4c129lnczad>
  + MCU CARD FOR TIVA TM4C129ENCZAD - <https://www.mikroe.com/mcu-card-for-tiva-tm4c129enczad>
  + MCU CARD FOR TIVA TM4C1299KCZAD - <https://www.mikroe.com/mcu-card-for-tiva-tm4c1299kczad>
  + MCU CARD FOR TIVA TM4C1299NCZAD - <https://www.mikroe.com/mcu-card-for-tiva-tm4c1299nczad>
  + MCU CARD 2 FOR TIVA TM4C123GH6PZ - <https://www.mikroe.com/mcu-card-2-for-tiva-tm4c123gh6pz>
  + MCU CARD 3 FOR TIVA TM4C1294KCPDT - <https://www.mikroe.com/mcu-card-3-for-tiva-tm4c1294kcpdt>
  + MCU CARD 3 FOR TIVA TM4C1294NCPDT - <https://www.mikroe.com/mcu-card-3-for-tiva-tm4c1294ncpdt>
  + MCU CARD 3 FOR TIVA TM4C129EKCPDT - <https://www.mikroe.com/mcu-card-3-for-tiva-tm4c129ekcpdt>
  + MCU CARD 3 FOR TIVA TM4C129ENCPDT - <https://www.mikroe.com/mcu-card-3-for-tiva-tm4c129encpdt>
  + EasyMx PRO v7 for Tiva MCU card with TM4C123GH6PMI - <https://www.mikroe.com/easymx-pro-v7-tiva-tm4c123gh6pmi>
  + EasyMx PRO v7 for Tiva MCU card with TM4C123GH6PZL - <https://www.mikroe.com/easymx-pro-v7-tiva-tm4c123gh6pzl>
  + EasyMx PRO v7 for Tiva MCU card with TM4C129XNCZAD - <https://www.mikroe.com/easymx-pro-v7-tiva-tm4c129xnczad>
+ Updated adequate manifest files
+ Configuration files updated

### Development kit package

+ Updated mikroSDK development kit to conform to newly added hardware stated in last paragraph
  + Added all 17 new hardware definitions

### Added adequate configuration files

Configuration files for following hardware has been added:

+ Clicker 2 for PIC18FJ - <https://www.mikroe.com/clicker-2-pic18fj>
+ PIC32MX Clicker - <https://www.mikroe.com/clicker-pic32mx>
+ Clicker 2 for PIC32MZ - <https://www.mikroe.com/clicker-2-for-pic32mz>

### PLANET DEBUG

PLANET DEBUG setups updated with mikroSDK packages.

Now, all click examples can be run out of the box on PLANET DEBUG.

3 systems have been updated:
  + Fusion for TIVA v8 - 1
  + Fusion for TIVA v8 - 2
  + Fusion for TIVA v8 - 3

For more information on PLANET DEBUG please visit following LINK - <https://www.mikroe.com/planet-debug>

### Recursive build automation

Updated powershell script for automated recursive SDK build.

Files changed:

+ ReadMe.md - <https://github.com/MikroElektronika/mikrosdk_v2/blob/master/scripts/ReadMe.md>
+ build-hal-recursion-ci.ps1 - <https://github.com/MikroElektronika/mikrosdk_v2/blob/master/scripts/build-hal-recursion-ci.ps1>
+ utility.ps1 - <https://github.com/MikroElektronika/mikrosdk_v2/blob/master/scripts/utility.ps1>
+ customParams.json - <https://github.com/MikroElektronika/mikrosdk_v2/blob/master/scripts/customParams.json>

For more information, please refer to main script DESCRIPTION:<https://github.com/MikroElektronika/mikrosdk_v2/blob/master/scripts/ReadMe.md)

### Test examples

Added following test examples:

+ clicks
  + SPI
    + click_magneto
  + UART
    + click_samltouch

Examples are used for testing mikroSDK specific module cases.

+++ scripts/build-hal-recursion-ci.ps1
@@ -7,7 +7,9 @@ param (
     [String]$coreSelect = $false,
     [Parameter(Mandatory=$false)]
     [String]$chipRegex = $false,
-    [Switch]$customParams = $false
+    [Switch]$customParams = $false,
+    [Switch]$examples = $false,
+    [Switch]$remove = $false
 )
 
 . '.\utility.ps1'
@@ -17,6 +19,8 @@ if ( $customParams ) {
     $silent = $customParamsContent.silent
     $Env:CORE = $customParamsContent.coreSelect
     $chipRegex = $customParamsContent.chipRegex
+    $examples = $customParamsContent.examples
+    $remove = $customParamsContent.remove
 }
 
 ## --------------------------------------
@@ -71,6 +75,7 @@ if ( $chipRegex -ne $false ) {
 } else {
     $defFiles = Get-ChildItem -Path $defsPath -Filter *.json
 }
+$Global:chipBuildCount = $defFiles.Count
 # -----------------------------------------------------------------------------
 
 # -----------------------------------------------------------------------------
@@ -117,8 +122,10 @@ foreach( $defFile in $defFiles ) {
 
         if ( $packageCount -eq 0 ) {
             $mcuName = $def.mcu
+            $package = ""
         } else {
             $mcuName = $def.mcu + $def.packages[$counter]
+            $package = $def.packages[$counter]
         }
 
         if ( $mcuDefinitions -contains $mcuName ) {
@@ -145,6 +152,14 @@ foreach( $defFile in $defFiles ) {
                     -Name "MSDK_HAL_LOW_LEVEL_TARGET" -Value "mikroe"
                 $configuration += Utils-Config `
                     -Name "MEMAKE_SDK_TYPE" -Value "normal"
+                if ( $examples ) {
+                    $configuration += Utils-Config `
+                        -Name "MSDK_TEST_EXAMPLES" -Value "true"
+                }
+
+                Utils-Generate-Cfg-File -mcu $mcuName -output $rootOutDir -package $package
+                $configuration += Utils-Config `
+                    -Name "MEMAKE_LINKER_CFG" -Value "$rootOutDir/$mcuName/CFG_DIR/$mcuName.json"
 
                 ## Configure project.
                 $memakeOutput = Memake-Configure -SrcDir $srcDir `
@@ -171,7 +186,9 @@ foreach( $defFile in $defFiles ) {
 
                 ## Log output
                 Utils-RecursionLog -Mcu $mcuName -LogFile $buildLogFile
-                Remove-Item -Path $outDir -Recurse
+                if ( $remove ) {
+                    Remove-Item -Path $outDir -Recurse
+                }
 
             }
